requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
scikit-learn
streamlit
matplotlib
seaborn
plotly